+++ .vscode/settings.json
@@ -16,5 +16,6 @@
   // Remove from file watching
   "files.watcherExclude": {
     "**/.fvm": true
-  }
+  },
+  "java.configuration.updateBuildConfiguration": "automatic"
 }
