+++ android/app/src/main/AndroidManifest.xml
@@ -1,8 +1,7 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.flutter_app">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android">
     <uses-permission android:name="android.permission.INTERNET" />
    <application
-        android:label="flutter_app"
+        android:label="Motis Mitfahr-App"
         android:name="${applicationName}"
         android:icon="@mipmap/ic_launcher">
         <activity
