+++ android/app/src/main/kotlin/com/motis/mitfahr_app/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.flutter_app
+package com.motis.mitfahr_app
 
 import io.flutter.embedding.android.FlutterActivity
