+++ lib/account/pages/write_review_page.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/account/models/profile.dart';
-import 'package:flutter_app/account/models/review.dart';
-import 'package:flutter_app/util/loading_button.dart';
-import 'package:flutter_app/util/profiles/profile_widget.dart';
-import 'package:flutter_app/util/profiles/reviews/custom_rating_bar.dart';
-import 'package:flutter_app/util/profiles/reviews/custom_rating_bar_size.dart';
-import 'package:flutter_app/util/supabase.dart';
+import 'package:motis_mitfahr_app/account/models/profile.dart';
+import 'package:motis_mitfahr_app/account/models/review.dart';
+import 'package:motis_mitfahr_app/util/loading_button.dart';
+import 'package:motis_mitfahr_app/util/profiles/profile_widget.dart';
+import 'package:motis_mitfahr_app/util/profiles/reviews/custom_rating_bar.dart';
+import 'package:motis_mitfahr_app/util/profiles/reviews/custom_rating_bar_size.dart';
+import 'package:motis_mitfahr_app/util/supabase.dart';
 import 'package:progress_state_button/progress_button.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
