+++ lib/drives/models/drive.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/account/models/profile.dart';
-import 'package:flutter_app/rides/models/ride.dart';
-import 'package:flutter_app/util/trip/trip.dart';
-import 'package:flutter_app/util/supabase.dart';
+import 'package:motis_mitfahr_app/account/models/profile.dart';
+import 'package:motis_mitfahr_app/rides/models/ride.dart';
+import 'package:motis_mitfahr_app/util/trip/trip.dart';
+import 'package:motis_mitfahr_app/util/supabase.dart';
 
 class Drive extends Trip {
   bool cancelled;
