+++ lib/drives/pages/drive_detail_page.dart
@@ -1,18 +1,18 @@
 import 'package:badges/badges.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_app/account/models/profile.dart';
-import 'package:flutter_app/drives/models/drive.dart';
-import 'package:flutter_app/rides/models/ride.dart';
-import 'package:flutter_app/util/big_button.dart';
-import 'package:flutter_app/util/custom_banner.dart';
-import 'package:flutter_app/util/custom_timeline_theme.dart';
-import 'package:flutter_app/util/icon_widget.dart';
-import 'package:flutter_app/util/locale_manager.dart';
-import 'package:flutter_app/util/own_theme_fields.dart';
-import 'package:flutter_app/util/profiles/profile_widget.dart';
-import 'package:flutter_app/util/profiles/profile_wrap_list.dart';
-import 'package:flutter_app/util/supabase.dart';
-import 'package:flutter_app/util/trip/trip_overview.dart';
+import 'package:motis_mitfahr_app/account/models/profile.dart';
+import 'package:motis_mitfahr_app/drives/models/drive.dart';
+import 'package:motis_mitfahr_app/rides/models/ride.dart';
+import 'package:motis_mitfahr_app/util/big_button.dart';
+import 'package:motis_mitfahr_app/util/custom_banner.dart';
+import 'package:motis_mitfahr_app/util/custom_timeline_theme.dart';
+import 'package:motis_mitfahr_app/util/icon_widget.dart';
+import 'package:motis_mitfahr_app/util/locale_manager.dart';
+import 'package:motis_mitfahr_app/util/own_theme_fields.dart';
+import 'package:motis_mitfahr_app/util/profiles/profile_widget.dart';
+import 'package:motis_mitfahr_app/util/profiles/profile_wrap_list.dart';
+import 'package:motis_mitfahr_app/util/supabase.dart';
+import 'package:motis_mitfahr_app/util/trip/trip_overview.dart';
 import 'package:timelines/timelines.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
