+++ lib/main.dart
@@ -1,13 +1,13 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_app/main_app.dart';
-import 'package:flutter_app/util/locale_manager.dart';
-import 'package:flutter_app/util/search/address_suggestion_manager.dart';
-import 'package:flutter_app/util/supabase.dart';
-import 'package:flutter_app/util/theme_manager.dart';
-import 'package:flutter_app/welcome/pages/reset_password_page.dart';
-import 'package:flutter_app/welcome/pages/welcome_page.dart';
+import 'package:motis_mitfahr_app/main_app.dart';
+import 'package:motis_mitfahr_app/util/locale_manager.dart';
+import 'package:motis_mitfahr_app/util/search/address_suggestion_manager.dart';
+import 'package:motis_mitfahr_app/util/supabase.dart';
+import 'package:motis_mitfahr_app/util/theme_manager.dart';
+import 'package:motis_mitfahr_app/welcome/pages/reset_password_page.dart';
+import 'package:motis_mitfahr_app/welcome/pages/welcome_page.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
