+++ lib/rides/models/ride.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/account/models/profile.dart';
-import 'package:flutter_app/util/trip/trip.dart';
-import 'package:flutter_app/util/supabase.dart';
+import 'package:motis_mitfahr_app/account/models/profile.dart';
+import 'package:motis_mitfahr_app/util/trip/trip.dart';
+import 'package:motis_mitfahr_app/util/supabase.dart';
 
 import '../../drives/models/drive.dart';
