+++ lib/rides/pages/ride_detail_page.dart
@@ -1,21 +1,21 @@
 import 'dart:math';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_app/account/models/profile.dart';
-import 'package:flutter_app/account/models/profile_feature.dart';
-import 'package:flutter_app/account/models/review.dart';
-import 'package:flutter_app/account/pages/write_review_page.dart';
-import 'package:flutter_app/drives/models/drive.dart';
-import 'package:flutter_app/rides/models/ride.dart';
-import 'package:flutter_app/util/big_button.dart';
-import 'package:flutter_app/util/custom_banner.dart';
-import 'package:flutter_app/util/profiles/reviews/custom_rating_bar_indicator.dart';
-import 'package:flutter_app/util/profiles/profile_widget.dart';
-import 'package:flutter_app/util/profiles/profile_wrap_list.dart';
-import 'package:flutter_app/util/profiles/reviews/custom_rating_bar_size.dart';
-import 'package:flutter_app/util/review_detail.dart';
-import 'package:flutter_app/util/supabase.dart';
-import 'package:flutter_app/util/trip/trip_overview.dart';
+import 'package:motis_mitfahr_app/account/models/profile.dart';
+import 'package:motis_mitfahr_app/account/models/profile_feature.dart';
+import 'package:motis_mitfahr_app/account/models/review.dart';
+import 'package:motis_mitfahr_app/account/pages/write_review_page.dart';
+import 'package:motis_mitfahr_app/drives/models/drive.dart';
+import 'package:motis_mitfahr_app/rides/models/ride.dart';
+import 'package:motis_mitfahr_app/util/big_button.dart';
+import 'package:motis_mitfahr_app/util/custom_banner.dart';
+import 'package:motis_mitfahr_app/util/profiles/reviews/custom_rating_bar_indicator.dart';
+import 'package:motis_mitfahr_app/util/profiles/profile_widget.dart';
+import 'package:motis_mitfahr_app/util/profiles/profile_wrap_list.dart';
+import 'package:motis_mitfahr_app/util/profiles/reviews/custom_rating_bar_size.dart';
+import 'package:motis_mitfahr_app/util/review_detail.dart';
+import 'package:motis_mitfahr_app/util/supabase.dart';
+import 'package:motis_mitfahr_app/util/trip/trip_overview.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 import '../../welcome/pages/login_page.dart';
