+++ lib/rides/pages/rides_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/rides/pages/search_ride_page.dart';
-import 'package:flutter_app/util/trip/trip_page_builder.dart';
+import 'package:motis_mitfahr_app/rides/pages/search_ride_page.dart';
+import 'package:motis_mitfahr_app/util/trip/trip_page_builder.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 import '../../util/supabase.dart';
