+++ lib/util/review_detail.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/util/locale_manager.dart';
-import 'package:flutter_app/util/profiles/profile_chip.dart';
-import 'package:flutter_app/util/profiles/reviews/custom_rating_bar_size.dart';
+import 'package:motis_mitfahr_app/util/locale_manager.dart';
+import 'package:motis_mitfahr_app/util/profiles/profile_chip.dart';
+import 'package:motis_mitfahr_app/util/profiles/reviews/custom_rating_bar_size.dart';
 
 import '../account/models/review.dart';
 import 'profiles/reviews/custom_rating_bar_indicator.dart';
