+++ lib/util/search/address_search_delegate.dart
@@ -1,7 +1,7 @@
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_app/util/search/address_suggestion_manager.dart';
-import 'package:flutter_app/util/search/address_suggestion.dart';
+import 'package:motis_mitfahr_app/util/search/address_suggestion_manager.dart';
+import 'package:motis_mitfahr_app/util/search/address_suggestion.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 class AddressSearchDelegate extends SearchDelegate<AddressSuggestion?> {
