+++ lib/util/search/address_search_field.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/util/search/address_search_delegate.dart';
-import 'package:flutter_app/util/search/address_suggestion.dart';
+import 'package:motis_mitfahr_app/util/search/address_search_delegate.dart';
+import 'package:motis_mitfahr_app/util/search/address_suggestion.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 class AddressSearchField extends StatelessWidget {
