+++ lib/util/theme_manager.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/util/storage_manager.dart';
+import 'package:motis_mitfahr_app/util/storage_manager.dart';
 import 'own_theme_fields.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
