+++ lib/util/trip/drive_card.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/util/custom_timeline_theme.dart';
+import 'package:motis_mitfahr_app/util/custom_timeline_theme.dart';
 
-import 'package:flutter_app/util/trip/trip_card.dart';
+import 'package:motis_mitfahr_app/util/trip/trip_card.dart';
 import '../../drives/models/drive.dart';
 import '../../drives/pages/drive_detail_page.dart';
 import 'package:timelines/timelines.dart';
