+++ lib/util/trip/trip_card.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/util/trip/trip.dart';
+import 'package:motis_mitfahr_app/util/trip/trip.dart';
 
 abstract class TripCard<T extends Trip> extends StatelessWidget {
   final T trip;
