+++ lib/util/trip/trip_overview.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/drives/models/drive.dart';
-import 'package:flutter_app/util/locale_manager.dart';
-import 'package:flutter_app/util/trip/trip.dart';
+import 'package:motis_mitfahr_app/drives/models/drive.dart';
+import 'package:motis_mitfahr_app/util/locale_manager.dart';
+import 'package:motis_mitfahr_app/util/trip/trip.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 class TripOverview extends StatelessWidget {
