+++ lib/util/trip/trip_page_builder.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/util/trip/trip.dart';
-import 'package:flutter_app/util/trip/trip_card.dart';
-import 'package:flutter_app/util/trip/trip_stream_builder.dart';
+import 'package:motis_mitfahr_app/util/trip/trip.dart';
+import 'package:motis_mitfahr_app/util/trip/trip_card.dart';
+import 'package:motis_mitfahr_app/util/trip/trip_stream_builder.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 class TripPageBuilder {
