+++ lib/util/trip/trip_stream_builder.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/util/trip/trip.dart';
+import 'package:motis_mitfahr_app/util/trip/trip.dart';
 
 import 'trip_card.dart';
