+++ lib/welcome/pages/forgot_password_page.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_app/util/email_field.dart';
-import 'package:flutter_app/util/loading_button.dart';
+import 'package:motis_mitfahr_app/util/email_field.dart';
+import 'package:motis_mitfahr_app/util/loading_button.dart';
 import 'package:progress_state_button/progress_button.dart';
-import 'package:flutter_app/util/supabase.dart';
+import 'package:motis_mitfahr_app/util/supabase.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 class ForgotPasswordPage extends StatefulWidget {
