+++ pubspec.yaml
@@ -1,5 +1,5 @@
-name: flutter_app
-description: A new Flutter project.
+name: motis_mitfahr_app
+description: A ride-sharing app by the Motis team.
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
